refactor: add new hook useAccordionFilter
